+++ Editor/Inspector/MergeSkinnedMeshEditor.cs
@@ -23,11 +23,11 @@ namespace Anatawa12.AvatarOptimizer
             };
         }
 
-        SerializedProperty _renderersSetProp;
-        SerializedProperty _staticRenderersSetProp;
-        SerializedProperty _removeEmptyRendererObjectProp;
-        SerializedProperty _skipEnablementMismatchedRenderers;
-        PrefabSafeSet.EditorUtil<Material> _doNotMergeMaterials;
+        SerializedProperty _renderersSetProp = null!; // initialized in OnEnable
+        SerializedProperty _staticRenderersSetProp = null!; // initialized in OnEnable
+        SerializedProperty _removeEmptyRendererObjectProp = null!; // initialized in OnEnable
+        SerializedProperty _skipEnablementMismatchedRenderers = null!; // initialized in OnEnable
+        PrefabSafeSet.EditorUtil<Material> _doNotMergeMaterials = null!; // initialized in OnEnable
 
         private void OnEnable()
         {
@@ -40,7 +40,7 @@ namespace Anatawa12.AvatarOptimizer
             _doNotMergeMaterials = PrefabSafeSet.EditorUtil<Material>.Create(
                 serializedObject.FindProperty("doNotMergeMaterials"),
                 nestCount,
-                x => x.objectReferenceValue as Material,
+                x => (Material)x.objectReferenceValue,
                 (x, v) => x.objectReferenceValue = v);
         }
 
@@ -68,7 +68,7 @@ namespace Anatawa12.AvatarOptimizer
 
         public void MergeMaterials(MergeSkinnedMesh merge)
         {
-            var materials = new HashSet<Material>();
+            var materials = new HashSet<Material?>();
             var renderersSetAsList = merge.renderersSet.GetAsList();
             var staticRenderersSetAsList = merge.staticRenderersSet.GetAsList();
             var ofRenderers = renderersSetAsList.Select(EditSkinnedMeshComponentUtil.GetMaterials);
@@ -77,6 +77,7 @@ namespace Anatawa12.AvatarOptimizer
                          .SelectMany((x, renderer) => x.Select((mat, material) => (mat, renderer, material)))
                          .GroupBy(x => x.mat))
             {
+                if (group.Key == null) continue;
                 materials.Add(group.Key);
                 if (group.Count() == 1)
                 {