+++ Editor/Processors/OriginalState.cs
@@ -14,7 +14,7 @@ namespace Anatawa12.AvatarOptimizer.Processors
         // avatar root => transform matrix (root.worldToLocalMatrix * transform.localToWorldMatrix)
         // we can local to world by root.localToWorldMatrix * thisMatrix
         private Dictionary<Transform, Matrix4x4> _originalTransforms = new Dictionary<Transform, Matrix4x4>();
-        public Transform AvatarRoot { get; set; }
+        public Transform AvatarRoot { get; set; } = null!; // set by FetchOriginalStatePass
 
         public void Register(Transform transform)
         {